feat(managed-v1): HerculesIAM provider, returnTo, 271f0ff previews
- Consolidate all IAM access UI into src/components/providers/
  hercules-iam.tsx (optimistic rendering, signed-out redirect grace,
  temporary-access retry with watchdog toast, /auth/* access routes);
  delete the four superseded files.
- Callback honors the signin({ returnTo }) OIDC state with an
  open-redirect guard (raw + decoded checks); RequireSignIn hands the
  attempted URL to the landing page via router state.
- Minimize .hercules/iam.jsonc to commented examples.
- Pin @usehercules/auth and @usehercules/convex to the 271f0ff previews,
  bump @usehercules/sdk to ^1.15.12, refresh the variant lockfile.
