+++ .github/template-variants/managed-v1/src/App.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from "react-router-dom";
 import { DefaultProviders } from "./components/providers/default.tsx";
-import { AuthAccessStatus } from "./pages/auth/AccessStatus.tsx";
+import { IamAccessRoute } from "./components/providers/hercules-iam.tsx";
 import AuthCallback from "./pages/auth/Callback.tsx";
 import Index from "./pages/Index.tsx";
 import NotFound from "./pages/NotFound.tsx";
@@ -11,7 +11,7 @@ export default function App() {
       <Routes>
         <Route path="/" element={<Index />} />
         <Route path="/auth/callback" element={<AuthCallback />} />
-        <Route path="/auth/*" element={<AuthAccessStatus />} />
+        <Route path="/auth/*" element={<IamAccessRoute />} />
         {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
         <Route path="*" element={<NotFound />} />
       </Routes>
